+++ src/repositories/enrollment/index.ts
@@ -1,0 +1,1 @@
+export * from "@/repositories/enrollment/enrollment.repository";