+++ src/repositories/student/index.ts
@@ -1,0 +1,1 @@
+export * from "@/repositories/student/student.repository";